Useful error message and usage for FSL script

+++ scripts/FSL.sh
@@ -4,10 +4,7 @@
 # boot.img is the concatenation of whoosh and the Linux Kernel
 # boot.imx is the final image (boot.img with i.MX Header)
 
-export PATH=$PATH:/home/alex/arm-2010q1/bin:
-export CROSS_COMPILE=arm-none-linux-gnueabi-
-
-USAGE="pouet"
+USAGE="$0 -k <kernel Image> -m <path to mkimage> [mmcdevice]"
 
 while getopts hk:m: OPT; do
     case "$OPT" in
@@ -31,22 +28,27 @@ done
 
 shift $(expr $OPTIND - 1)
 
-if [ $# -eq 0 ]; then
-    echo $USAGE >&2
-    exit 1
+DEVICE=$1
+
+if [ -z "$KERNEL" -o -z "$MKIMAGE" ]; then
+	echo $USAGE
+	exit 1
 fi
 
 # Make sure we are not trying to write to the primary partition
-if [ $1 = "/dev/sda" ]; then
+if [ "$1" = "/dev/sda" ]; then
 	echo "ERROR: Forbidden to write to /dev/sda"
-	echo "Syntax:  $0 [device]"
-	echo "Example: $0 /dev/sdg"
+	echo $USAGE
 	exit 1
 fi
 
 # Build the bootloader
 make clean
 make
+if [ $? -ne 0 ]; then
+	echo "Compilation failed"
+	exit 1
+fi
 
 # Dynamic calculation of the start address
 mkdir out
@@ -58,6 +60,9 @@ START_ADDR=`echo "obase=16; $TEMP_START_ADDR" | bc`
 
 # Generate the images and copy it to the sd card
 ${MKIMAGE} -n $(dirname ${MKIMAGE})/imximage.cfg -T imximage -e $START_ADDR -d out/boot.img out/boot.imx
-sudo dd if=out/boot.imx of=$1 bs=512 seek=2
-sudo sync
+
+if [ -n "$DEVICE" ]; then
+	sudo dd if=out/boot.imx of=$1 bs=512 seek=2
+	sudo sync
+fi
